Use selection’s “changed” signal rather than tree’s “cursor-changed”
In GTK 3.4.4 and 3.6.1, the “cursor-changed” signal is triggered when the
tree is destroyed, which meant get_selection() returned None and the callback
crashed. This behaviour seems to be gone again in GTK 3.8.2.

Workaround inspired by “Changes in ‘cursor-changed’ signal of Gtk.TreeView
cause issues with netzob”, https://dev.netzob.org/issues/234.